コマンドライン引数の解析を別ファイルに
リファクタリング
